# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel API Documentation",
        "description": "Документация API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/shared-wishlists/{id}": {
            "get": {
                "tags": [
                    "Shared wishlists"
                ],
                "summary": "Список желаний или список дел для гостя",
                "description": "Список дел (type = todo) открывается, только если владелец включил доступ по ссылке; у его позиций isSelected означает «выполнено», а поля jointGift нет. Отмечать дела гость может, только если guestsCanCheck = true. У позиций списка желаний есть поле jointGift: данные совместного подарка или null",
                "operationId": "25a622f31374dd4c709bcc09f807f597",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "ulid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список желаний",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "format": "ulid"
                                                },
                                                "type": {
                                                    "type": "string",
                                                    "enum": [
                                                        "gift",
                                                        "todo"
                                                    ]
                                                },
                                                "title": {
                                                    "type": "string"
                                                },
                                                "color": {
                                                    "description": "Ключ цвета фона списка",
                                                    "type": "string",
                                                    "example": "lavender",
                                                    "enum": [
                                                        "white",
                                                        "lavender",
                                                        "pink",
                                                        "peach",
                                                        "lemon",
                                                        "mint",
                                                        "sky"
                                                    ]
                                                },
                                                "guestsCanCheck": {
                                                    "description": "Может ли гость отмечать дела выполненными (POST /items/check); у списка желаний false",
                                                    "type": "boolean"
                                                },
                                                "guestNameRequired": {
                                                    "description": "Должен ли гость указать имя, отмечая дела; false, если отмечать нельзя",
                                                    "type": "boolean"
                                                },
                                                "dueDate": {
                                                    "description": "Срок списка дел или дата события списка желаний (Y-m-d); null — дата не указана",
                                                    "type": "string",
                                                    "format": "date",
                                                    "nullable": true
                                                },
                                                "username": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "avatarUrl": {
                                                    "description": "Фотография владельца списка; null — показывается первая буква имени",
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "items": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "string",
                                                                "format": "ulid"
                                                            },
                                                            "isSelected": {
                                                                "type": "boolean"
                                                            },
                                                            "label": {
                                                                "type": "string"
                                                            },
                                                            "urls": {
                                                                "type": "array",
                                                                "items": {
                                                                    "type": "string"
                                                                }
                                                            },
                                                            "priority": {
                                                                "type": "integer",
                                                                "nullable": true
                                                            },
                                                            "price": {
                                                                "type": "integer",
                                                                "nullable": true
                                                            },
                                                            "checkedBy": {
                                                                "oneOf": [
                                                                    {
                                                                        "$ref": "#/components/schemas/TodoItemCheckedBy"
                                                                    }
                                                                ],
                                                                "nullable": true
                                                            },
                                                            "jointGift": {
                                                                "oneOf": [
                                                                    {
                                                                        "$ref": "#/components/schemas/SharedWishlistJointGift"
                                                                    }
                                                                ],
                                                                "nullable": true
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Список не найден, является заметкой или списком дел без доступа по ссылке"
                    },
                    "429": {
                        "description": "Слишком много запросов"
                    }
                }
            }
        },
        "/api/v1/shared-wishlists/{wishlist}/items": {
            "patch": {
                "tags": [
                    "Shared wishlists"
                ],
                "summary": "Выбрать подарки и получить токен брони",
                "operationId": "746f33989c28895b4a326e5632a866fa",
                "parameters": [
                    {
                        "name": "wishlist",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "ulid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "item_ids"
                                ],
                                "properties": {
                                    "item_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "ulid"
                                        }
                                    },
                                    "joint_gifts": {
                                        "description": "Совместные подарки; item_id каждого должен входить в item_ids",
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "item_id",
                                                "name"
                                            ],
                                            "properties": {
                                                "item_id": {
                                                    "type": "string",
                                                    "format": "ulid"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "maxLength": 50
                                                },
                                                "contact": {
                                                    "type": "string",
                                                    "nullable": true,
                                                    "maxLength": 100
                                                },
                                                "comment": {
                                                    "type": "string",
                                                    "nullable": true,
                                                    "maxLength": 200
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Позиции выбраны. В data.reservation — токен для отмены выбора",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "reservation": {
                                                    "$ref": "#/components/schemas/SharedWishlistReservation"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Позиция уже выбрана другим гостем или совместный подарок указан для невыбранной позиции"
                    },
                    "429": {
                        "description": "Слишком много запросов"
                    }
                }
            }
        },
        "/api/v1/shared-wishlists/{wishlist}/items/check": {
            "post": {
                "tags": [
                    "Shared wishlists"
                ],
                "summary": "Отметить дела выполненными",
                "description": "Только для списка дел с доступом по ссылке, где владелец разрешил гостям отмечать дела (guestsCanCheck). Гость только ставит отметку, снять её может лишь владелец. Уже выполненные дела не изменяются, в том числе не меняется имя того, кто их отметил",
                "operationId": "76a19235b47692c253563f6cd6cf0f14",
                "parameters": [
                    {
                        "name": "wishlist",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "ulid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "item_ids"
                                ],
                                "properties": {
                                    "item_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "ulid"
                                        }
                                    },
                                    "name": {
                                        "description": "Имя гостя для подписи под делами. Обязательно, если guestNameRequired = true",
                                        "type": "string",
                                        "nullable": true,
                                        "maxLength": 50
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Дела отмечены; в ответе актуальный список, как в GET /api/v1/shared-wishlists/{id}"
                    },
                    "404": {
                        "description": "Список не найден или гостям не разрешено отмечать в нём дела"
                    },
                    "422": {
                        "description": "Некоторых дел нет в этом списке или не указано обязательное имя"
                    },
                    "429": {
                        "description": "Слишком много запросов"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "SharedWishlistJointGift": {
                "description": "Совместный подарок: данные организатора для других гостей. В ответы владельцу списка не попадает",
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Иван, коллега Ани",
                        "maxLength": 50
                    },
                    "contact": {
                        "description": "Контакт организатора; выводится обычным текстом, без ссылки",
                        "type": "string",
                        "example": "@ivan_k",
                        "nullable": true,
                        "maxLength": 100
                    },
                    "comment": {
                        "type": "string",
                        "example": "Собираем до 10 октября",
                        "nullable": true,
                        "maxLength": 200
                    }
                },
                "type": "object"
            },
            "TodoItemCheckedBy": {
                "description": "Кто отметил дело выполненным; есть только у позиций списка дел, у невыполненного дела null. guest = false — отметил владелец, guest = true — гость по ссылке, name — его имя или null, если гость имя не указал",
                "properties": {
                    "guest": {
                        "type": "boolean"
                    },
                    "name": {
                        "type": "string",
                        "nullable": true,
                        "maxLength": 50
                    }
                },
                "type": "object"
            }
        }
    },
...
}